[S06P31D205-108]feat : :sparkles: 리뷰 통계 정보 세팅

+++ AOS/app/src/main/java/com/ssafy/moabang/data/model/response/ReviewStatResponse.kt
@@ -12,5 +12,5 @@ data class ReviewStatResponse(
     val r_activity: String,
     val r_clearTime: Int,
     val r_recPlayer: Int,
-    val r_hint: Int,
+    val r_hint: Float,
 ): Parcelable
